requirements.txt
alembic==1.7.7
atomicwrites==1.4.0
attrs==21.4.0
click==8.1.2
colorama==0.4.4
config==0.5.1
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.8
pytest==7.1.1
python-dotenv==0.20.0
SQLAlchemy==1.4.35
tomli==2.0.1
Werkzeug==2.1.1
zipp==3.8.0
